Move Tree component under tree package/folder
